requirements.txt
datasets==2.16
huggingface-hub==0.34
evaluate
jiwer
transformers==4.54.0
librosa
torchcodec==0.3
